+++ tests/PhpWord/Tests/Writer/Word2007/Part/DocumentTest.php
@@ -135,7 +135,11 @@ class DocumentTest extends \PHPUnit_Framework_TestCase
 
         $tabs = array(new \PhpOffice\PhpWord\Style\Tab('right', 9090));
         $phpWord = new PhpWord();
-        $phpWord->addParagraphStyle('pStyle', array('align' => 'center', 'tabs' => $tabs)); // Style #1
+        $phpWord->addParagraphStyle('pStyle', array(
+            'align' => 'center',
+            'tabs' => $tabs,
+            'shading' => array('fill' => 'FFFF99'),
+        )); // Style #1
         $phpWord->addFontStyle('fStyle', array('size' => '20', 'bold' => true, 'allCaps' => true,
             'scale' => 200, 'spacing' => 240, 'kerning' => 10)); // Style #2
         $phpWord->addTitleStyle(1, array('color' => '333333', 'doubleStrikethrough' => true)); // Style #3